+++ CHANGELOG.md
@@ -3,6 +3,16 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
+## [0.14.0] - 2022-03-19
+### Added
+- `Select` component
+### Breaking
+- `Toggle` `state` prop renamed to `value`
+### Changed
+- added support for `disabled` prop for `Toggle`
+### Fixed
+- missing `Stats` export
+
 ## [0.13.0] - 2022-02-23
 ### Added
 - `Spacer` util component